requirements.txt
blinker==1.5
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.4
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==0.20.0
pytz==2022.5
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.42
Unidecode==1.3.6
urllib3==1.26.9
Werkzeug==2.2.2
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
